// src/app/editor/components/editor/Text2Col/types.ts
export type TText2ColData = {
  title: string;
  titleColor: "#3F28C6";
  subtitle: string;
  description: string;
  images: Array<Record<string, string>>;
  editorImages: Array<Record<string, string>>;
  content: string;
  sortIndex: number;
  secondEditorImages: Array<Record<string, string>>;
  secondContent: string;
};